Split the GameRenderer class into a package with 3 classes. Also added comments.
