+++ src/test/java/entities/PlayerTest.java
@@ -1,6 +1,5 @@
 package entities;
 
-import client.GameRenderer;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
